+++ src/renderer/components/wallet/assets/AssetsNav.tsx
@@ -49,11 +49,6 @@ export const AssetsNav: React.FC = (): JSX.Element => {
           key: MenuKey.HISTORY,
           label: intl.formatMessage({ id: 'common.history' }),
           path: walletRoutes.history.path()
-        },
-        {
-          key: MenuKey.WALLETSETTINGS,
-          label: intl.formatMessage({ id: 'common.settings' }),
-          path: walletRoutes.walletSettings.path()
         }
       ] as MenuType[],
     [intl]
@@ -63,7 +58,6 @@ export const AssetsNav: React.FC = (): JSX.Element => {
   const poolSharesRoute = matchPath(walletRoutes.poolShares.path(), pathname)
   const bondsRoute = matchPath(walletRoutes.bonds.path(), pathname)
   const matchHistoryRoute = matchPath(walletRoutes.history.path(), pathname)
-  const matchWalletSettingsRoute = matchPath(walletRoutes.walletSettings.path(), pathname)
 
   const activeMenu: MenuKey = useMemo(() => {
     if (assetsRoute) {
@@ -74,12 +68,10 @@ export const AssetsNav: React.FC = (): JSX.Element => {
       return MenuKey.BONDS
     } else if (matchHistoryRoute) {
       return MenuKey.HISTORY
-    } else if (matchWalletSettingsRoute) {
-      return MenuKey.WALLETSETTINGS
     } else {
       return MenuKey.UNKNOWN
     }
-  }, [assetsRoute, poolSharesRoute, bondsRoute, matchHistoryRoute, matchWalletSettingsRoute])
+  }, [assetsRoute, poolSharesRoute, bondsRoute, matchHistoryRoute])
 
   return (
     <>