+++ src/renderer/components/wallet/settings/index.ts
@@ -1,1 +1,2 @@
 export { WalletSettings } from './WalletSettings'
+export { UnlockWalletSettings } from './UnlockWalletSettings'